Permit missing chef_config in ArtifactoryCookbookSource
When `policyfile/artifactory_cookbook_source` is included in tests,
it hooks rspec suite execution to initialize/install the policyfile
before  execution the tests.

In test cases, this means a valid `chef_config` is not present because
the hook does not provid a valid `chef_config`.

This change ensures that even when chef-config is absent, we fall back
correctly to the environment variable `ARTIFACTORY_API_KEY`

Fixes #138

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>